+++ Client/Game/Enviroment/BiomeManagement/Biomes/Grasslands/Hills.cs
@@ -15,7 +15,7 @@ namespace VoxelValley.Client.Game.Enviroment.BiomeManagement.Biomes
             return GenerationUtilities.MapToWorldByte(GenerationUtilities.FBMPerlin(x, z, 5, 2, 0.4f));
         }
 
-        internal override int GetVoxel(int x, int y, int z, int height)
+        internal override ushort GetVoxel(int x, int y, int z, int height)
         {
 
             if (y > height)
